move grades dialog into `d2l-activity-grades` folder
