+++ components/d2l-activity-editor/d2l-activity-grades/d2l-activity-grades-dialog.js
@@ -1,16 +1,16 @@
-import './d2l-activity-grades/d2l-activity-grade-candidate-selector';
-import './d2l-activity-grades/d2l-activity-grade-category-selector';
+import './d2l-activity-grade-candidate-selector';
+import './d2l-activity-grade-category-selector';
 import '@brightspace-ui/core/components/button/button.js';
 import '@brightspace-ui/core/components/dialog/dialog.js';
 import '@brightspace-ui/core/components/icons/icon.js';
 import '@brightspace-ui/core/components/inputs/input-radio-spacer.js';
-import { sharedAssociateGrade as associateGradeStore, shared as store } from './state/activity-store.js';
+import { sharedAssociateGrade as associateGradeStore, shared as store } from '../state/activity-store.js';
 import { css, html } from 'lit-element/lit-element';
-import { ActivityEditorWorkingCopyDialogMixin } from './mixins/d2l-activity-editor-working-copy-dialog-mixin.js';
+import { ActivityEditorWorkingCopyDialogMixin } from '../mixins/d2l-activity-editor-working-copy-dialog-mixin.js';
 import { bodySmallStyles } from '@brightspace-ui/core/components/typography/styles.js';
 import { formatNumber } from '@brightspace-ui/intl/lib/number.js';
 import { GradebookStatus } from 'siren-sdk/src/activities/associateGrade/AssociateGradeEntity.js';
-import { LocalizeActivityEditorMixin } from './mixins/d2l-activity-editor-lang-mixin.js';
+import { LocalizeActivityEditorMixin } from '../mixins/d2l-activity-editor-lang-mixin.js';
 import { MobxLitElement } from '@adobe/lit-mobx';
 import { radioStyles } from '@brightspace-ui/core/components/inputs/input-radio-styles.js';
 import { RtlMixin } from '@brightspace-ui/core/mixins/rtl-mixin.js';
